Restructure mac module into apple module
