+++ src/apple/ios/mod.rs
@@ -1,0 +1,8 @@
+//
+// Sysinfo
+//
+// Copyright (c) 2021 Guillaume Gomez
+//
+
+mod component;
+pub use self::component::Component;
