+++ src/apple/processor.rs
